+++ quartus/stock/vexriscv/sw/evaluation/benchmark.h
@@ -2,14 +2,14 @@
 
 #define ELEMENTS 4
 
-void soft_vmul(uint32_t arr[ELEMENTS], uint32_t a[ELEMENTS], uint32_t b[ELEMENTS]);
+void soft_vmul(int8_t arr[ELEMENTS], int8_t a[ELEMENTS], int8_t b[ELEMENTS]);
 
-uint32_t soft_vacc(uint32_t a[ELEMENTS]);
+int32_t soft_vacc(int8_t a[ELEMENTS]);
 
-uint32_t soft_vmaxe(uint32_t a[ELEMENTS]);
+int8_t soft_vmaxe(int8_t a[ELEMENTS]);
 
-uint32_t soft_vmine(uint32_t a[ELEMENTS]);
+int8_t soft_vmine(int8_t a[ELEMENTS]);
 
-void soft_vmaxx(uint32_t arr[ELEMENTS], uint32_t a[ELEMENTS], uint32_t s);
+void soft_vmaxx(int8_t arr[ELEMENTS], int8_t a[ELEMENTS], int8_t s);
 
-void soft_vsrli(uint32_t arr[ELEMENTS], uint32_t a[ELEMENTS], uint32_t s);
+void soft_vsrli(int8_t arr[ELEMENTS], int8_t a[ELEMENTS], int8_t s);
